backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langchain>=0.3.7
langchain-groq>=0.2.2
langchain-community>=0.3.7
langchain-core>=0.3.27
pinecone>=5.0.0
sentence-transformers>=3.0.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
requests>=2.31.0
sse-starlette>=2.1.0
pydantic>=2.8.0
pypdf>=4.3.0
langchain-text-splitters>=0.3.0
